requirements.txt
flask
requests
python-jose
cryptography>=42.0.0